+++ packages/theme-default/__tests__/code-command.test.ts
@@ -1,12 +1,12 @@
 import { describe, expect, it } from 'vitest'
-import { COMMAND_RE, getCommand } from '../src/markdown/commands'
+import { COMMAND_RE, getCommands } from '../src/markdown/commands'
 
 describe('code commands', () => {
   it('highlight', () => {
     expect(COMMAND_RE.test('const foo = 1 // [svp! hl]')).toBeTruthy()
     expect(COMMAND_RE.test('const foo = 1 // [svp! hl:10]')).toBeTruthy()
 
-    expect(COMMAND_RE.exec('const foo = 1 // [svp! hl]')).toMatchInlineSnapshot(`
+    expect(COMMAND_RE.exec('const foo = 1 // [svp! hl] // [svp! ++]')).toMatchInlineSnapshot(`
       [
         "// [svp! hl]",
         "hl",
@@ -20,13 +20,6 @@ describe('code commands', () => {
         undefined,
       ]
     `)
-
-    expect(getCommand('const foo = 1 // [svp! hl]')).toMatchInlineSnapshot(`
-      [
-        "hl",
-        "const foo = 1 ",
-      ]
-    `)
   })
 
   it('diff', () => {
@@ -36,22 +29,39 @@ describe('code commands', () => {
     expect(COMMAND_RE.test('const foo = 1 // [svp! ++]')).toBeTruthy()
     expect(COMMAND_RE.test('const foo = 1 // [svp! --]')).toBeTruthy()
 
-    expect(getCommand('const foo = 1 // [svp! ++]')).toMatchInlineSnapshot(`
-      [
-        "++",
-        "const foo = 1 ",
-      ]
-    `)
+    // expect(getCommands('const foo = 1 // [svp! ++]')).toMatchInlineSnapshot(`
+    //   [
+    //     [
+    //       "++",
+    //     ],
+    //     "const foo = 1 ",
+    //   ]
+    // `)
   })
 
   it('focus', () => {
     expect(COMMAND_RE.test('const foo = 1 // [svp! fc]')).toBeTruthy()
     expect(COMMAND_RE.test('const foo = 1 // [svp! !!]')).toBeTruthy()
 
-    expect(getCommand('const foo = 1 // [svp! !!:3]')).toMatchInlineSnapshot(`
+    // expect(getCommands('const foo = 1 // [svp! !!:3]')).toMatchInlineSnapshot(`
+    //   [
+    //     [
+    //       "!!:3",
+    //     ],
+    //     "const foo = 1 ",
+    //   ]
+    // `)
+  })
+
+  it('multi command', async () => {
+    expect(getCommands('const foo = 1 // [svp! ++] // [svp! !!:3] // [svp! ~~]')).toMatchInlineSnapshot(`
       [
-        "!!:3",
-        "const foo = 1 ",
+        [
+          "++",
+          "!!:3",
+          "~~",
+        ],
+        "const foo = 1   ",
       ]
     `)
   })